+++ plugins/source/aws/resources/services/dynamodb/table_continuous_backups.go
@@ -1,7 +1,10 @@
 package dynamodb
 
 import (
+	"context"
+
 	"github.com/apache/arrow/go/v14/arrow"
+	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
@@ -26,4 +29,26 @@ func tableContinuousBackups() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchDynamodbTableContinuousBackups(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	par := parent.Item.(*types.TableDescription)
+
+	cl := meta.(*client.Client)
+	svc := cl.Services(client.AWSServiceDynamodb).Dynamodb
+
+	output, err := svc.DescribeContinuousBackups(ctx, &dynamodb.DescribeContinuousBackupsInput{
+		TableName: par.TableName,
+	}, func(options *dynamodb.Options) {
+		options.Region = cl.Region
+	})
+	if err != nil {
+		if cl.IsNotFoundError(err) {
+			return nil
+		}
+		return err
+	}
+
+	res <- output.ContinuousBackupsDescription
+	return nil
 }
